requirements.txt
alembic==1.15.1
python-dotenv==1.1.0
python-telegram-bot==22.0
pytz==2025.2
SQLAlchemy==2.0.40
